[Rust SDK][refactor]: Renamed wasmedge as wasmedge-sdk.
