+++ bindings/rust/Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["wasmedge-sys", "wasmedge-types", "wasmedge"]
+members = ["wasmedge-sys", "wasmedge-types", "wasmedge-sdk"]
 exclude = ["build/", "utils/"]
