+++ bindings/rust/wasmedge-sdk/Cargo.toml
@@ -1,12 +1,12 @@
 [package]
-name = "wasmedge"
+name = "wasmedge-sdk"
 version = "0.1.0"
 edition = "2021"
 description = "WasmEdge Runtime is a high-performance, extensible, and hardware optimized WebAssembly Virtual Machine for automotive, cloud, AI, and blockchain applications."
 license = "Apache-2.0"
 readme = "README.md"
 documentation = "https://docs.rs/crate/wasmedge-sdk"
-repository = "https://github.com/WasmEdge/WasmEdge/bindings/rust/wasmedge"
+repository = "https://github.com/WasmEdge/WasmEdge/bindings/rust/wasmedge-sdk"
 categories = ["api-bindings", "wasm"]
 exclude = ["examples/"]
