+++ bindings/rust/wasmedge-sdk/examples/call_func_in_threads.rs
@@ -7,7 +7,7 @@ use std::{
     sync::{Arc, Mutex},
     thread,
 };
-use wasmedge::{
+use wasmedge_sdk::{
     config::{CommonConfigOptions, ConfigBuilder},
     params, Engine, Executor, Module, Statistics, Store, WasmVal,
 };
