+++ bindings/rust/wasmedge-sdk/src/compiler.rs
