+++ bindings/rust/wasmedge-sdk/src/externals/global.rs
