+++ bindings/rust/wasmedge-sdk/src/externals/memory.rs
