+++ bindings/rust/wasmedge-sdk/src/instance.rs
