+++ bindings/rust/wasmedge-sdk/src/store.rs
