+++ bindings/rust/wasmedge-sdk/src/vm.rs
